Reorganized and checked that both BLE inteface and UART interface are working. Need to look into proper include structures, static data_array[] behavior and doing intensive processing in callback functions.
